+++ sdk/netapp/arm-netapp/src/models/parameters.ts
@@ -51,9 +51,6 @@ import {
   BackupVaultPatch as BackupVaultPatchMapper,
   BackupRestoreFiles as BackupRestoreFilesMapper,
   BackupsMigrationRequest as BackupsMigrationRequestMapper,
-  Bucket as BucketMapper,
-  BucketPatch as BucketPatchMapper,
-  BucketCredentialsExpiry as BucketCredentialsExpiryMapper,
 } from "../models/mappers.js";
 
 export const accept: OperationParameter = {
@@ -83,7 +80,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2025-01-01-preview",
+    defaultValue: "2025-03-01",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -606,37 +603,4 @@ export const body38: OperationParameter = {
 export const body39: OperationParameter = {
   parameterPath: "body",
   mapper: BackupsMigrationRequestMapper,
-};
-
-export const bucketName: OperationURLParameter = {
-  parameterPath: "bucketName",
-  mapper: {
-    constraints: {
-      Pattern: new RegExp(
-        "^([a-z]|(\\d(?!\\d{0,2}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3})))([a-z\\d]|(\\.(?!(\\.|-)))|(-(?!\\.))){1,61}[a-z\\d]$",
-      ),
-      MaxLength: 63,
-      MinLength: 3,
-    },
-    serializedName: "bucketName",
-    required: true,
-    type: {
-      name: "String",
-    },
-  },
-};
-
-export const body40: OperationParameter = {
-  parameterPath: "body",
-  mapper: BucketMapper,
-};
-
-export const body41: OperationParameter = {
-  parameterPath: ["options", "body"],
-  mapper: BucketPatchMapper,
-};
-
-export const body42: OperationParameter = {
-  parameterPath: "body",
-  mapper: BucketCredentialsExpiryMapper,
 };