+++ sdk/netapp/arm-netapp/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release History
     
+## 22.0.0 (2025-07-15)
+    
+### Features Added
+  - Interface SubscriptionQuotaItemList has a new optional parameter nextLink
+  - Added Type Alias NetAppResourceQuotaLimitsListNextResponse
+  - Enum KnownReplicationSchedule has a new value TenMinutely
+### Breaking Changes
+  - Enum KnownReplicationSchedule no longer has value "10Minutely"
+
+    
+    
 ## 21.6.0-beta.1 (2025-05-14)
 Compared with version 21.5.0